+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- `Styra Link: Bundle Update` command added.
+- Info badges added to top of extension's home page in VS Code.
+
 ## [1.0.1] - 2023-03-16
 
 ### Fixed