+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-04-06
+
 ### Added
 
 - `Styra Link: Validate Compliance` command added.
@@ -115,7 +117,8 @@ whether there is anything to overwrite, then asks that first, mirroring the CLI 
 - `Styra Link: Config Git` command added.
 
 <!-- Be sure to add an entry here for each release! -->
-[unreleased]: https://github.com/StyraInc/vscode-styra/compare/v1.0.1-alpha...HEAD
+[unreleased]: https://github.com/StyraInc/vscode-styra/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/StyraInc/vscode-styra/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/StyraInc/vscode-styra/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/StyraInc/vscode-styra/compare/v0.0.6-alpha...v1.0.0
 [0.0.6-alpha]: https://github.com/StyraInc/vscode-styra/compare/v0.0.5-alpha...v0.0.6-alpha